Variant Allele productisation
new X flags on variant callers
--Xmin-vac (default 0)
--Xmin-vaf (default 0.0)

When one of them is not 0 it triggers VA handling.

+++ src/com/rtg/vcf/annotation/VariantAllelicFractionAnnotation.java
@@ -70,7 +70,7 @@ public class VariantAllelicFractionAnnotation extends AbstractDerivedFormatAnnot
   }
 
   private double sum(final double... a) {
-    int s = 0;
+    double s = 0;
     for (final double v : a) {
       s += v;
     }
